requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
aiosqlite==0.20.0
alembic==1.13.1
anyio==4.9.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.1.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
connexion==3.0.6
cryptography==45.0.4
flake8==7.0.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pluggy==1.6.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.25.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.37.2
swagger_ui_bundle==1.1.0
typing_extensions==4.14.0
urllib3==2.2.1
uvicorn==0.34.0
websockets==12.0
Werkzeug==3.1.3
